examples/ollama/requirements.txt
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
asttokens==3.0.0
astunparse==1.6.3
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
claudette==0.1.0
cosette==0.0.4
cryptography==44.0.0
decorator==5.1.1
distro==1.9.0
docutils==0.21.2
execnb==0.1.11
executing==2.1.0
fastcore==1.7.28
fewsats==0.0.9
ghapi==1.0.6
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
ipython==8.31.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jiter==0.8.2
keyring==25.6.0
llms-txt==0.0.4
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
msglm==0.0.4
nbdev==2.3.34
nh3==0.2.20
openai==1.58.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pkginfo==1.12.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
sherlock-domains==0.0.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
toolslm==0.1.0
tqdm==4.67.1
traitlets==5.14.3
twine==6.0.1
typing_extensions==4.12.2
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
zipp==3.21.0
